% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_model.R
\name{covariate_meta}
\alias{covariate_meta}
\title{Capture the encoding of a covariate so it can be reproduced on new data}
\usage{
covariate_meta(covariate)
}
\arguments{
\item{covariate}{A vector (numeric, factor, or character).}
}
\value{
A list describing how to build the design matrix for this covariate.
}
\description{
Capture the encoding of a covariate so it can be reproduced on new data
}
\keyword{internal}
